refactor(go): consolidate currency utilities
